Fix ensuring that collection is selected
