+++ frontend/src/metabase/common/components/DashboardPicker/components/DashboardPicker.tsx
@@ -1,12 +1,6 @@
 import type { Ref } from "react";
-import {
-  forwardRef,
-  useCallback,
-  useEffect,
-  useImperativeHandle,
-  useMemo,
-} from "react";
-import { useDeepCompareEffect, useLatest } from "react-use";
+import { forwardRef, useCallback, useImperativeHandle, useMemo } from "react";
+import { useDeepCompareEffect } from "react-use";
 
 import {
   skipToken,
@@ -19,6 +13,7 @@ import { useSelector } from "metabase/lib/redux";
 import { getUserPersonalCollectionId } from "metabase/selectors/user";
 import type { CollectionItemModel, Dashboard } from "metabase-types/api";
 
+import { useEnsureCollectionSelected } from "../../CollectionPicker";
 import { CollectionItemPickerResolver } from "../../CollectionPicker/components/CollectionItemPickerResolver";
 import { getPathLevelForItem } from "../../CollectionPicker/utils";
 import { LoadingSpinner, NestedItemPicker } from "../../EntityPicker";
@@ -203,16 +198,13 @@ const DashboardPickerInner = (
     [currentCollection, userPersonalCollectionId, onPathChange],
   );
 
-  const onItemSelectRef = useLatest(onItemSelect); // use ref to prevent effect from running too often
-
-  useEffect(
-    function ensureFolderSelected() {
-      if (!pathProp && defaultPath[0].selectedItem) {
-        onItemSelectRef.current(defaultPath[0].selectedItem);
-      }
-    },
-    [pathProp, defaultPath, onItemSelectRef],
-  );
+  useEnsureCollectionSelected({
+    currentCollection,
+    enabled: path === defaultPath,
+    options,
+    useRootCollection: initialValue?.id == null,
+    onItemSelect,
+  });
 
   if (error) {
     return <LoadingAndErrorWrapper error={error} />;
